// zoetrope-editor/src/lib/renderer/types.ts
export type PathSegment =
  | { type: 'move-to'; x: number; y: number }
  | { type: 'line-to'; x: number; y: number }
  | { type: 'curve-to'; x: number; y: number; c1x: number; c1y: number; c2x: number; c2y: number }
  | { type: 'close-path' }